requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
reportlab==4.0.7
python-barcode==0.15.1
Pillow==10.1.0
jinja2==3.1.2
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
python-dateutil==2.8.2 